동시성 문제 해결
