+++ src/main/java/com/review/rsproject/repository/ReviewRepository.java
@@ -2,11 +2,13 @@ package com.review.rsproject.repository;
 
 import com.review.rsproject.domain.Review;
 import com.review.rsproject.dto.response.ReviewCountDto;
+import jakarta.persistence.LockModeType;
 import jakarta.persistence.TypedQuery;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Lock;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
@@ -15,6 +17,8 @@ import java.util.Optional;
 
 public interface ReviewRepository extends JpaRepository<Review, Long> {
 
+
+    @Lock(LockModeType.OPTIMISTIC)
     @Query("select new com.review.rsproject.dto.response.ReviewCountDto(count(r), sum(r.star)) from Review r where r.platform.id = :platformId")
     ReviewCountDto findByStar(@Param("platformId") Long platformId);
